refactor Repeated
