+++ pkg/specs/enum_test.go
@@ -1,28 +1,23 @@
-package compare
+package specs
 
-import (
-	"testing"
+import "testing"
 
-	"github.com/jexia/semaphore/pkg/specs"
-)
-
-func TestCompareEnumValues(t *testing.T) {
-	newValue := func() *specs.EnumValue {
-		return &specs.EnumValue{
+func TestEnumValueCompare(t *testing.T) {
+	newValue := func() *EnumValue {
+		return &EnumValue{
 			Key:      "a",
 			Position: 1,
 		}
 	}
 
-	shouldBeEqual := func(t *testing.T, expected, given *specs.EnumValue) {
-		got := CompareEnumValues(expected, given)
-		if got != nil {
-			t.Fatalf("CompareEnumValues() returns unexpected error: %v", got)
+	shouldBeEqual := func(t *testing.T, expected, given *EnumValue) {
+		if err := given.Compare(expected); err != nil {
+			t.Fatalf("CompareEnumValues() returns unexpected error: %v", err)
 		}
 	}
 
-	shouldNotBeEqual := func(t *testing.T, expected, given *specs.EnumValue) {
-		if CompareEnumValues(expected, given) == nil {
+	shouldNotBeEqual := func(t *testing.T, expected, given *EnumValue) {
+		if given.Compare(expected) == nil {
 			t.Fatalf("CompareEnumValues() = nil, want an error")
 		}
 	}
@@ -59,37 +54,37 @@ func TestCompareEnumValues(t *testing.T) {
 	})
 }
 
-func TestCompareEnums(t *testing.T) {
-	createEnum := func() *specs.Enum {
-		valA := &specs.EnumValue{
+func TestEnumCompare(t *testing.T) {
+	createEnum := func() *Enum {
+		valA := &EnumValue{
 			Key:         "a",
 			Position:    0,
 			Description: "",
 		}
 
-		valB := &specs.EnumValue{
+		valB := &EnumValue{
 			Key:         "b",
 			Position:    1,
 			Description: "",
 		}
 
-		return &specs.Enum{
+		return &Enum{
 			Name:        "letters",
 			Description: "",
-			Keys:        map[string]*specs.EnumValue{valA.Key: valA, valB.Key: valB},
-			Positions:   map[int32]*specs.EnumValue{valA.Position: valA, valB.Position: valB},
+			Keys:        map[string]*EnumValue{valA.Key: valA, valB.Key: valB},
+			Positions:   map[int32]*EnumValue{valA.Position: valA, valB.Position: valB},
 		}
 	}
 
-	shouldBeEqual := func(t *testing.T, expected, given *specs.Enum) {
-		got := CompareEnums(expected, given)
+	shouldBeEqual := func(t *testing.T, expected, given *Enum) {
+		got := given.Compare(expected)
 		if got != nil {
 			t.Fatalf("CompareEnums() returns unexpected error: %v", got)
 		}
 	}
 
-	shouldNotBeEqual := func(t *testing.T, expected, given *specs.Enum) {
-		if CompareEnums(expected, given) == nil {
+	shouldNotBeEqual := func(t *testing.T, expected, given *Enum) {
+		if given.Compare(expected) == nil {
 			t.Fatalf("CompareEnums() = nil, want an error")
 		}
 	}
@@ -117,18 +112,18 @@ func TestCompareEnums(t *testing.T) {
 	})
 
 	t.Run("nil and not nil enums should not be equal", func(t *testing.T) {
-		shouldNotBeEqual(t, nil, &specs.Enum{})
-		shouldNotBeEqual(t, &specs.Enum{}, nil)
+		shouldNotBeEqual(t, nil, &Enum{})
+		shouldNotBeEqual(t, &Enum{}, nil)
 	})
 
 	t.Run("enums with nil items under Keys/Positions should be equal", func(t *testing.T) {
 		enum, another := createEnum(), createEnum()
 
-		enum.Keys = map[string]*specs.EnumValue{"a": nil, "b": nil}
-		enum.Positions = map[int32]*specs.EnumValue{0: nil, 1: nil}
+		enum.Keys = map[string]*EnumValue{"a": nil, "b": nil}
+		enum.Positions = map[int32]*EnumValue{0: nil, 1: nil}
 
-		another.Keys = map[string]*specs.EnumValue{"a": nil, "b": nil}
-		another.Positions = map[int32]*specs.EnumValue{0: nil, 1: nil}
+		another.Keys = map[string]*EnumValue{"a": nil, "b": nil}
+		another.Positions = map[int32]*EnumValue{0: nil, 1: nil}
 
 		shouldBeEqual(t, enum, another)
 	})
@@ -144,7 +139,7 @@ func TestCompareEnums(t *testing.T) {
 	t.Run("enums with different size of Keys should not be equal", func(t *testing.T) {
 		enum, another := createEnum(), createEnum()
 
-		another.Keys["c"] = &specs.EnumValue{}
+		another.Keys["c"] = &EnumValue{}
 
 		shouldNotBeEqual(t, enum, another)
 	})
@@ -152,7 +147,7 @@ func TestCompareEnums(t *testing.T) {
 	t.Run("enums with different size of Positions should not be equal", func(t *testing.T) {
 		enum, another := createEnum(), createEnum()
 
-		another.Positions[3] = &specs.EnumValue{}
+		another.Positions[3] = &EnumValue{}
 
 		shouldNotBeEqual(t, enum, another)
 	})
@@ -160,7 +155,7 @@ func TestCompareEnums(t *testing.T) {
 	t.Run("enums with different values under positions should not be equal", func(t *testing.T) {
 		enum, another := createEnum(), createEnum()
 
-		another.Positions[0] = &specs.EnumValue{}
+		another.Positions[0] = &EnumValue{}
 
 		shouldNotBeEqual(t, enum, another)
 	})
@@ -168,7 +163,7 @@ func TestCompareEnums(t *testing.T) {
 	t.Run("enums with different values under keys should not be equal", func(t *testing.T) {
 		enum, another := createEnum(), createEnum()
 
-		another.Keys["a"] = &specs.EnumValue{}
+		another.Keys["a"] = &EnumValue{}
 
 		shouldNotBeEqual(t, enum, another)
 	})
