+++ pkg/specs/message_test.go
@@ -1,38 +1,37 @@
-package compare
+package specs
 
 import (
 	"testing"
 
-	"github.com/jexia/semaphore/pkg/specs"
 	"github.com/jexia/semaphore/pkg/specs/labels"
 	"github.com/jexia/semaphore/pkg/specs/types"
 )
 
 func TestCompareMessages(t *testing.T) {
-	createScalar := func() *specs.Property {
-		return &specs.Property{
+	createScalar := func() *Property {
+		return &Property{
 			Name:     "age",
 			Path:     "dog",
 			Position: 0,
 			Label:    labels.Required,
-			Template: specs.Template{
-				Scalar: &specs.Scalar{
+			Template: Template{
+				Scalar: &Scalar{
 					Type: types.Int32,
 				},
 			},
 		}
 	}
-	messageA := specs.Message{
+	messageA := Message{
 		"age": createScalar(),
 	}
-	messageB := specs.Message{
+	messageB := Message{
 		"age":   createScalar(),
 		"color": createScalar(),
 	}
 
 	type args struct {
-		expected specs.Message
-		given    specs.Message
+		expected Message
+		given    Message
 	}
 	tests := []struct {
 		name    string
@@ -71,7 +70,7 @@ func TestCompareMessages(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			if err := CompareMessages(tt.args.expected, tt.args.given); (err != nil) != tt.wantErr {
+			if err := tt.args.given.Compare(tt.args.expected); (err != nil) != tt.wantErr {
 				t.Errorf("CompareMessages() error = %v, wantErr %v", err, tt.wantErr)
 			}
 		})
