+++ pkg/specs/property.go
@@ -1,9 +1,7 @@
 package specs
 
 import (
-	"encoding/json"
-	"errors"
-	"sort"
+	"fmt"
 
 	"github.com/jexia/semaphore/pkg/specs/labels"
 	"github.com/jexia/semaphore/pkg/specs/metadata"
@@ -38,242 +36,6 @@ type Expression interface {
 	Position() string
 }
 
-// Scalar value.
-type Scalar struct {
-	Default interface{} `json:"default,omitempty"`
-	Type    types.Type  `json:"type,omitempty"`
-}
-
-// UnmarshalJSON corrects the 64bit data types in accordance with golang
-func (scalar *Scalar) UnmarshalJSON(data []byte) error {
-	if scalar == nil {
-		return nil
-	}
-
-	type sc Scalar
-	t := sc{}
-
-	err := json.Unmarshal(data, &t)
-	if err != nil {
-		return err
-	}
-
-	*scalar = Scalar(t)
-	scalar.Clean()
-
-	return nil
-}
-
-// Clean fixes the type casting issue of unmarshal
-func (scalar *Scalar) Clean() {
-	if scalar.Default == nil {
-		return
-	}
-
-	switch scalar.Type {
-	case types.Int64, types.Sint64, types.Sfixed64:
-		_, ok := scalar.Default.(int64)
-		if !ok {
-			scalar.Default = int64(scalar.Default.(float64))
-		}
-	case types.Uint64, types.Fixed64:
-		_, ok := scalar.Default.(uint64)
-		if !ok {
-			scalar.Default = uint64(scalar.Default.(float64))
-		}
-	case types.Int32, types.Sint32, types.Sfixed32:
-		_, ok := scalar.Default.(int32)
-		if !ok {
-			scalar.Default = int32(scalar.Default.(float64))
-		}
-	case types.Uint32, types.Fixed32:
-		_, ok := scalar.Default.(uint32)
-		if !ok {
-			scalar.Default = uint32(scalar.Default.(float64))
-		}
-	}
-}
-
-// Clone scalar value.
-func (scalar Scalar) Clone() *Scalar {
-	return &Scalar{
-		Default: scalar.Default,
-		Type:    scalar.Type,
-	}
-}
-
-// Enum represents a enum configuration
-type Enum struct {
-	*metadata.Meta
-	Name        string                `json:"name,omitempty"`
-	Description string                `json:"description,omitempty"`
-	Keys        map[string]*EnumValue `json:"keys,omitempty"`
-	Positions   map[int32]*EnumValue  `json:"positions,omitempty"`
-}
-
-// Clone enum schema.
-func (e Enum) Clone() *Enum { return &e }
-
-// EnumValue represents a enum configuration
-type EnumValue struct {
-	*metadata.Meta
-	Key         string `json:"key,omitempty"`
-	Position    int32  `json:"position,omitempty"`
-	Description string `json:"description,omitempty"`
-}
-
-// Repeated represents an array type of fixed size.
-type Repeated []Template
-
-// Template returns a Template for a given array. It checks the types of internal
-// elements to detect the data type(s).
-func (repeated Repeated) Template() (Template, error) {
-	var (
-		template Template
-		dataType types.Type
-		position int
-	)
-
-	// check if all element types are the same
-	// TODO: remove once "oneOf" support is added
-	for position, template = range repeated {
-		if position == 0 {
-			dataType = template.Type()
-
-			continue
-		}
-
-		if template.Type() != dataType {
-			return template, errors.New("not implemented: one of")
-		}
-	}
-
-	// get rid of default value if scalar type
-	if template.Scalar != nil {
-		template = template.Clone()
-		template.Scalar.Default = nil
-	}
-
-	return template, nil
-}
-
-// Clone repeated.
-func (repeated Repeated) Clone() Repeated {
-	var clone = make([]Template, len(repeated))
-
-	for index, template := range repeated {
-		clone[index] = template.Clone()
-	}
-
-	return clone
-}
-
-// Message represents an object which keeps the original order of keys.
-type Message map[string]*Property
-
-// SortedProperties returns the available properties as a properties list
-// ordered base on the properties position.
-func (message Message) SortedProperties() PropertyList {
-	result := make(PropertyList, 0, len(message))
-
-	for _, property := range message {
-		result = append(result, property)
-	}
-
-	sort.Sort(result)
-	return result
-}
-
-// Clone the message.
-func (message Message) Clone() Message {
-	var clone = make(map[string]*Property, len(message))
-
-	for key := range message {
-		clone[key] = message[key].Clone()
-	}
-
-	return clone
-}
-
-// Template contains property schema. This is a union type (Only one field must be set).
-type Template struct {
-	Reference *PropertyReference `json:"reference,omitempty"` // Reference represents a property reference made inside the given property
-
-	// Only one of the following fields should be set
-	Scalar   *Scalar  `json:"scalar,omitempty"`
-	Enum     *Enum    `json:"enum,omitempty"`
-	Repeated Repeated `json:"repeated,omitempty"`
-	Message  Message  `json:"message,omitempty"`
-}
-
-// Type returns the type of the given template.
-func (t Template) Type() types.Type {
-	if t.Message != nil {
-		return types.Message
-	}
-
-	if t.Repeated != nil {
-		return types.Array
-	}
-
-	if t.Enum != nil {
-		return types.Enum
-	}
-
-	if t.Scalar != nil {
-		return t.Scalar.Type
-	}
-
-	return types.Unknown
-}
-
-// Clone internal value.
-func (t Template) Clone() Template {
-	var clone = Template{
-		Reference: t.Reference.Clone(),
-	}
-
-	if t.Scalar != nil {
-		clone.Scalar = t.Scalar.Clone()
-	}
-
-	if t.Enum != nil {
-		clone.Enum = t.Enum.Clone()
-	}
-
-	if t.Repeated != nil {
-		clone.Repeated = t.Repeated.Clone()
-	}
-
-	if t.Message != nil {
-		clone.Message = t.Message.Clone()
-	}
-
-	return clone
-}
-
-// PropertyList represents a list of properties
-type PropertyList []*Property
-
-func (list PropertyList) Len() int           { return len(list) }
-func (list PropertyList) Swap(i, j int)      { list[i], list[j] = list[j], list[i] }
-func (list PropertyList) Less(i, j int) bool { return list[i].Position < list[j].Position }
-
-// Get attempts to return a property inside the given list with the given name
-func (list PropertyList) Get(key string) *Property {
-	for _, item := range list {
-		if item == nil {
-			continue
-		}
-
-		if item.Name == key {
-			return item
-		}
-	}
-
-	return nil
-}
-
 // Property represents a value property.
 type Property struct {
 	*metadata.Meta
@@ -294,8 +56,8 @@ type Property struct {
 }
 
 // DefaultValue returns rge default value for a given property.
-func (prop *Property) DefaultValue() interface{} {
-	t := prop.Template
+func (property *Property) DefaultValue() interface{} {
+	t := property.Template
 	switch {
 	case t.Scalar != nil:
 		return t.Scalar.Default
@@ -311,30 +73,60 @@ func (prop *Property) DefaultValue() interface{} {
 }
 
 // Empty checks if the property has any defined type
-func (prop *Property) Empty() bool {
-	return prop.Template.Type() == types.Unknown
+func (property *Property) Empty() bool {
+	return property.Template.Type() == types.Unknown
 }
 
 // Clone makes a deep clone of the given property
-func (prop *Property) Clone() *Property {
-	if prop == nil {
+func (property *Property) Clone() *Property {
+	if property == nil {
 		return &Property{}
 	}
 
 	return &Property{
-		Meta:        prop.Meta,
-		Position:    prop.Position,
-		Description: prop.Description,
-		Name:        prop.Name,
-		Path:        prop.Path,
+		Meta:        property.Meta,
+		Position:    property.Position,
+		Description: property.Description,
+		Name:        property.Name,
+		Path:        property.Path,
 
-		Expr:    prop.Expr,
-		Raw:     prop.Raw,
-		Options: prop.Options,
-		Label:   prop.Label,
+		Expr:    property.Expr,
+		Raw:     property.Raw,
+		Options: property.Options,
+		Label:   property.Label,
 
-		Template: prop.Template.Clone(),
+		Template: property.Template.Clone(),
 	}
+}
+
+// Compare checks the given property against the provided one.
+// TODO: return error stack
+func (property *Property) Compare(expected *Property) error {
+	if expected == nil {
+		return fmt.Errorf("unable to check types for '%s' no schema given", property.Path)
+	}
+
+	if property.Template.Type() != expected.Type() {
+		return fmt.Errorf("cannot use type (%s) for '%s', expected (%s)", property.Type(), property.Path, expected.Type())
+	}
+
+	if property.Label != expected.Label {
+		return fmt.Errorf("cannot use label (%s) for '%s', expected (%s)", property.Label, property.Path, expected.Label)
+	}
+
+	if !property.Empty() && expected.Empty() {
+		return fmt.Errorf("property '%s' has a nested object but schema does not '%s'", property.Path, expected.Name)
+	}
+
+	if !expected.Empty() && property.Empty() {
+		return fmt.Errorf("schema '%s' has a nested object but property does not '%s'", expected.Name, property.Path)
+	}
+
+	if err := property.Template.Compare(expected.Template); err != nil {
+		return fmt.Errorf("nested schema mismatch under property '%s': %w", property.Path, err)
+	}
+
+	return nil
 }
 
 // ParameterMap is the initial map of parameter names (keys) and their (templated) values (values)
@@ -365,8 +157,8 @@ func (parameters *ParameterMap) Clone() *ParameterMap {
 		Property: parameters.Property.Clone(),
 	}
 
-	for key, prop := range parameters.Params {
-		result.Params[key] = prop.Clone()
+	for key, property := range parameters.Params {
+		result.Params[key] = property.Clone()
 	}
 
 	for key, value := range parameters.Options {
@@ -377,8 +169,8 @@ func (parameters *ParameterMap) Clone() *ParameterMap {
 		result.Header[key] = value.Clone()
 	}
 
-	for key, prop := range parameters.Stack {
-		result.Stack[key] = prop.Clone()
+	for key, property := range parameters.Stack {
+		result.Stack[key] = property.Clone()
 	}
 
 	return result
